circleci: separate build artifacts from build data.
The opencv build generates both framework and intemediate build
directory. Previously both were zipped together, creating a huge zip
file containing everything.

A simple solution would be to just zip the framework and upload it.
But if there's an error you'de want to debug the build, and since the
build is very long (40 mins), we can just treat the build directory as
an artifacts and upload it as well.